requirements.txt
tensorflow>=2.13.0
numpy>=1.22.0
Pillow>=9.0.0
Flask>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
